[installer] Move `personalAccessTokenSigningKey` out of experimental
Fixes #15028

This PR moves `personalAccessTokenSigningKey` out of `experimental`
and into `v1.Config` while marking it as `required` by default, with
a default secret name of `personal-access-token-signing-key`. It's
the environment's responsibility to make sure this secret exists for
Gitpod to work. (See #15028 and slack thread for more info on this).

This PR also adds
- validation checks for `personalAccessTokenSigningKey` secret
  along with the key
- depreciation check for existing config

Signed-off-by: Tarun Pothulapati <tarun@gitpod.io>